implement dry run
